REGLAGE DE PLUSIEURS DETAILS
- Suppression de l'ancienne classe terrain.h
- La classe partie.h a été renommée en terrain.h
- La sauvegarde du terrain après la partie a été mise en commentaire (pas ce qu'il faut faire)
- Les vectors ont été transformés en double

+++ vector.h
@@ -10,18 +10,18 @@ class point;
 class vector {
 public:
   vector();
-  vector(int x,int y);
+  vector(double x,double y);
   vector(const point& p,const point& q);
 
-  int x() const;
-  int y() const;
+  double x() const;
+  double y() const;
 
-  void change(int x, int y);
+  void change(double x, double y);
 
   void print(std::ostream& ost) const;
   void read(std::istream& ist);
 private:
-  int d_x,d_y;
+  double d_x,d_y;
 };
 
 }
